refactor: userId 중복검증 훅 및 submit 로직 리팩토링 #733
- 이전 구조는 useSubmitGuardWithConfirm 훅을 통한 파생 상태 관리였으나,
    버튼 클릭 시 에러 메시지 출력이 헷갈리고 확장성도 낮음
- useSubmitGuardWithConfirm 기반 파생 상태 구조 제거
- 중복확인 체크를 위한 userIdDuplicateChecked 상태와 제출 시도여부를 위한 submitAttempted 상태 도입
- 입력 예외 검증은 파생 상태, 중복 검증/시도 여부는 원본 상태로 분리
- handleSubmit에서 submitAttempted 기반 중복 체크 후 제출 처리
